feat: Further restructured the project.
