+++ src/Chrono.Tests/Chrono.Tests.csproj
@@ -29,6 +29,11 @@
     <Content Include="config.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
+    <None Remove="config.Local.json"/>
+    <Content Include="config.Local.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
   </ItemGroup>
+
 
 </Project>
