+++ src/Chrono.Tests/UnitTests/01_TaskLists.cs
@@ -1,4 +1,5 @@
 using Chrono.Entities;
+using Chrono.Features.TaskLists;
 using FluentAssertions;
 using NUnit.Framework;
 using Task = Chrono.Entities.Task;
