--- src/Chrono/Features/Categories/CategoryConfiguration.cs
@@ -1,18 +1,0 @@
-using Chrono.Entities;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace Chrono.Features.Categories;
-
-public class CategoryConfiguration : IEntityTypeConfiguration<Category>
-{
-    public void Configure(EntityTypeBuilder<Category> builder)
-    {
-        builder.Property(x => x.Name)
-            .HasMaxLength(100)
-            .IsRequired();
-
-        builder.Navigation(x => x.CreatedBy)
-            .AutoInclude();
-    }
-}
