+++ src/Chrono/Features/TaskLists/DeleteTaskList.cs
@@ -10,20 +10,6 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Chrono.Features.TaskLists;
 
-[Authorize] [Route("api/tasklists")]
-public class DeleteTaskListsController : ApiControllerBase
-{
-    [HttpDelete("{id:int}")]
-    [ProducesDefaultResponseType]
-    [ProducesResponseType(StatusCodes.Status204NoContent)]
-    [ProducesResponseType(StatusCodes.Status403Forbidden)]
-    public async Task<IActionResult> Delete(int id)
-    {
-        await Mediator.Send(new DeleteTaskList(id));
-        return NoContent();
-    }
-}
-
 public record DeleteTaskList(int Id) : IRequest;
 
 public class DeleteTaskListHandler : IRequestHandler<DeleteTaskList>
@@ -57,5 +43,19 @@ public class DeleteTaskListHandler : IRequestHandler<DeleteTaskList>
         _context.TaskLists.Remove(entity);
 
         await _context.SaveChangesAsync(cancellationToken);
+    }
+}
+
+[Authorize] [Route("api/tasklists")]
+public class DeleteTaskListsController : ApiControllerBase
+{
+    [HttpDelete("{id:int}")]
+    [ProducesDefaultResponseType]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
+    public async Task<IActionResult> Delete(int id)
+    {
+        await Mediator.Send(new DeleteTaskList(id));
+        return NoContent();
     }
 }
