--- src/Chrono/Features/Users/UserConfiguration.cs
@@ -1,27 +1,0 @@
-using Chrono.Entities;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace Chrono.Features.Users;
-
-public class UserConfiguration : IEntityTypeConfiguration<User>
-{
-    public void Configure(EntityTypeBuilder<User> builder)
-    {
-        builder.HasIndex(x => x.UserId)
-            .IsUnique();
-
-        builder.Property(x => x.UserId)
-            .IsRequired();
-
-        builder.Property(x => x.Name)
-            .IsRequired();
-
-        builder.HasOne(x => x.UserSettings)
-            .WithOne(x => x.User)
-            .HasPrincipalKey<User>(x => x.Id)
-            .HasForeignKey<UserSettings>(x => x.UserId)
-            .OnDelete(DeleteBehavior.Cascade)
-            .IsRequired(false);
-    }
-}
