--- src/Chrono/Features/Users/UserSettingsConfiguration.cs
@@ -1,14 +1,0 @@
-﻿using Chrono.Entities;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace Chrono.Features.Users;
-
-public class UserSettingsConfiguration : IEntityTypeConfiguration<UserSettings>
-{
-    public void Configure(EntityTypeBuilder<UserSettings> builder)
-    {
-        builder.Navigation(x => x.DefaultTaskList)
-            .AutoInclude();
-    }
-}
